add types for ts users
Previously, to use a type such as JavaColor for an argument to a
function in ts, it would quickly become unreadable as you would have
to use things such as InstanceType to obtain the type.
This update exposes all the types currently implemented.
All the exposed types are prefixed with a "T" to denote they are a type.
This change will only affect those who are using CTA with ts